pyproject.toml
[project]
name = "cell-classification"
dynamic = ["version"]
description = "A cell classifier using information limited to images of DAPI stained nuclei."
readme = "README.md"
license = {file = "LICENSE"}
authors = [
  {name = "Aurimas Greicius"},
  {email = "aurimasgreicius@yahoo.com"},
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: napari",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
]

# Python version requirement
requires-python = "==3.12.2"

# Pip dependencies
dependencies = [
    "pyqt5",
    "numpy",
    "magicgui",
    "qtpy",
    "scikit-image>=0.22.0",
    "jsonargparse",
    "ipykernel",  # For Jupyter notebook support
    "joblib",     # For parallel runs, random forests
    "pandas",
    "pytorch-lightning==2.3.3",
    "matplotlib",
    "tensorboard==2.17.1",
    "scikit-learn==1.5.1",
    "scipy==1.11.4",
    "lightning==2.3.3",
    "numba==0.60.0",
    "napari"
]


[project.optional-dependencies]
testing = [
    "tox",
    "pytest",  # https://docs.pytest.org/en/latest/contents.html
    "pytest-cov",  # https://pytest-cov.readthedocs.io/en/latest/
    "pytest-qt"  # https://pytest-qt.readthedocs.io/en/latest/
]

[project.entry-points."console_scripts"]
main = "nucleus_3d_classification.main:main"

[project.entry-points."napari.manifest"]
cell-classification = "nucleus_3d_classification:napari.yaml"



[build-system]
requires = ["setuptools>=42.0.0", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]


[tool.setuptools_scm]
write_to = "src/nucleus_3d_classification/_version.py"


[tool.black]
line-length = 79
target-version = ['py38', 'py39', 'py310']

[tool.ruff]
line-length = 79
lint.select = [
    "E", "F", "W", #flake8
    "UP", # pyupgrade
    "I", # isort
    "BLE", # flake8-blind-exception
    "B", # flake8-bugbear
    "A", # flake8-builtins
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "G", # flake8-logging-format
    "PIE", # flake8-pie
    "SIM", # flake8-simplify
]
lint.ignore = [
    "E501", # line too long. let black handle this
    "UP006", "UP007", # type annotation. As using magicgui require runtime type annotation then we disable this.
    "SIM117", # flake8-simplify - some of merged with statements are not looking great with black, reanble after drop python 3.9
]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".mypy_cache",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "*vendored*",
    "*_vendor*",
]

target-version = "py38"
fix = true
